Tried adding folders
But ended up dropping them because they aren't going to work quite as
intended. Basically you can't add an empty one.. you have to have
something in it for it to work. Also move the repo methods into its own
class so we can instantiate it and not have to pass the repo object
around.